examples/private_shapley/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "private-shapley"
version = "0.1.0"
description = "Private Shapley Value computation with coalition management and Merkle proofs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flwr-datasets[vision]>=0.5.0",
    "flwr[simulation]>=1.18.0",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "web3>=7.10.0",
    "eth-account>=0.13.6",
    "python-dotenv>=1.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "trizelabs"

[tool.flwr.app.components]
serverapp = "private_shapley.server_app:app"
clientapp = "private_shapley.client_app:app"

[tool.flwr.app.config]
num-server-rounds = 3
fraction-fit = 0.5
fraction-evaluate = 1.0
min-available-clients = 2
local-epochs = 1
learning-rate = 0.005
batch-size = 32
verbose = false
metrics-storage-path = 'logs'

[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 10

[tool.eth.account]
mnemonic = "$PRIVATE_SHAPLEY_MNEMONIC"

[tool.web3]
url = "http://127.0.0.1:8545"
contract_address = "$PRIVATE_SHAPLEY_CONTRACT"